SADXGG verification added
